requirements/development.txt
Werkzeug==0.6.2
django-extensions==0.6
django-rosetta==0.6.0
